# catalog-service/src/main/openapi/catalog-service-openapi.yml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: 'http://localhost:8080'
    description: Generated server url
...
paths:
  /api/category:
    post:
      tags:
        - CategoryController
      operationId: postApiCategory
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                string:
                  type: string
              x-examples:
                example-1:
                  string: string
      responses:
        '201':
          description: Created
...